+++ packages/spacecat-shared-data-access/src/models/suggestion/suggestion.model.js
@@ -34,6 +34,7 @@ class Suggestion extends BaseModel {
     ERROR: 'ERROR',
     OUTDATED: 'OUTDATED',
     PENDING_VALIDATION: 'PENDING_VALIDATION',
+    REJECTED: 'REJECTED',
   };
 
   static TYPES = {
@@ -110,7 +111,7 @@ class Suggestion extends BaseModel {
     }
   }
 
-  // add your customized method  here
+  // add your customized method here
 }
 
 export default Suggestion;